+++ src/lib/Layer.svelte
@@ -13,6 +13,11 @@ The Layer component needs to be placed inside a svelte-konva Stage component.
 </Stage>
 ```
 
+### Static config:
+By default svelte-konva will automatically update your config prop on `dragend` and `transformend` events to match the config state (position, rotation, scale, ...) with the internal Konva state. 
+If you additionally bind the config prop your reactive blocks will also be triggered once this happens. 
+There might be cases where this behavior is not beneficial in this case you can disable it by passing the `staticConfig = true` prop to the component.
+
 Further information: [Konva API docs](https://konvajs.org/api/Konva.Layer.html), [svelte-konva docs](https://konvajs.org/docs/svelte)
 -->
 <script lang="ts">
@@ -21,10 +26,11 @@ Further information: [Konva API docs](https://konvajs.org/api/Konva.Layer.html),
 	import { type Writable, writable } from 'svelte/store';
 	import { Container, getParentStage, setContainerContext } from '$lib/util/manageContext';
 	import { registerEvents } from '$lib/util/events';
-	import { copyExistingKeys } from './util/copy';
+	import { copyExistingKeys } from './util/object';
 
 	export let config: Konva.LayerConfig = {};
 	export let handle: Konva.Layer = new Konva.Layer(config);
+	export let staticConfig = false;
 
 	let inner = writable<null | Konva.Layer>(null);
 
@@ -41,15 +47,17 @@ Further information: [Konva API docs](https://konvajs.org/api/Konva.Layer.html),
 	onMount(() => {
 		$parent!.add(handle);
 
-		handle.on('transformend', () => {
-			copyExistingKeys(config, handle.getAttrs());
-			config = config;
-		});
+		if (!staticConfig) {
+			handle.on('transformend', () => {
+				copyExistingKeys(config, handle.getAttrs());
+				config = config;
+			});
 
-		handle.on('dragend', () => {
-			copyExistingKeys(config, handle.getAttrs());
-			config = config;
-		});
+			handle.on('dragend', () => {
+				copyExistingKeys(config, handle.getAttrs());
+				config = config;
+			});
+		}
 
 		registerEvents(dispatcher, handle);