+++ src/lib/Stage.svelte
@@ -10,6 +10,11 @@ Under the hood it creates a div element where the html canvas is attached to.
 </Stage>
 ```
 
+### Static config:
+By default svelte-konva will automatically update your config prop on `dragend` and `transformend` events to match the config state (position, rotation, scale, ...) with the internal Konva state. 
+If you additionally bind the config prop your reactive blocks will also be triggered once this happens. 
+There might be cases where this behavior is not beneficial in this case you can disable it by passing the `staticConfig = true` prop to the component.
+
 Further information: [Konva API docs](https://konvajs.org/api/Konva.Stage.html), [svelte-konva docs](https://konvajs.org/docs/svelte)
 -->
 <script lang="ts">
@@ -18,10 +23,11 @@ Further information: [Konva API docs](https://konvajs.org/api/Konva.Stage.html),
 	import { writable } from 'svelte/store';
 	import { registerEvents } from '$lib/util/events';
 	import { Container, setContainerContext } from '$lib/util/manageContext';
-	import { copyExistingKeys } from './util/copy';
+	import { copyExistingKeys } from './util/object';
 
 	export let config: Konva.ContainerConfig;
 	export let handle: null | Konva.Stage = null;
+	export let staticConfig = false;
 
 	let inner = writable<null | Konva.Stage>(null);
 
@@ -41,10 +47,12 @@ Further information: [Konva API docs](https://konvajs.org/api/Konva.Stage.html),
 			...config
 		});
 
-		handle.on('dragend', () => {
-			copyExistingKeys(config, handle!.getAttrs());
-			config = config;
-		});
+		if (!staticConfig) {
+			handle.on('dragend', () => {
+				copyExistingKeys(config, handle!.getAttrs());
+				config = config;
+			});
+		}
 
 		registerEvents(dispatcher, handle);