+++ src/tests/layer.test.ts
@@ -104,7 +104,7 @@ test('Correctly updates bound config on dragend', () => {
 	});
 
 	const component = rendered.component.$$;
-	const handle: Konva.Label = component.ctx[component.props['handle'] as number];
+	const handle: Konva.Layer = component.ctx[component.props['handle'] as number];
 
 	const stage = handle.getStage()!;
 	const rectangle = new Konva.Rect({ x: 0, y: 0, width: 100, height: 100, fill: 'red' });
@@ -120,6 +120,37 @@ test('Correctly updates bound config on dragend', () => {
 	const config = component.ctx[component.props['config'] as number];
 
 	expect(config).toStrictEqual({ ...CONFIG, x: 50 });
+});
+
+test('Does not update config if instantiated with staticConfig prop', async () => {
+	const CONFIG = { x: 0, draggable: true };
+	const oldConfig = { ...CONFIG };
+	const div = document.createElement('div');
+	const rendered = render(Layer, {
+		context: createMockParentContext(Container.Stage, div),
+		props: {
+			config: CONFIG,
+			staticConfig: true
+		}
+	});
+
+	const component = rendered.component.$$;
+	const handle: Konva.Layer = component.ctx[component.props['handle'] as number];
+
+	const stage = handle.getStage()!;
+	const rectangle = new Konva.Rect({ x: 0, y: 0, width: 100, height: 100, fill: 'red' });
+
+	handle.add(rectangle);
+
+	stage.draw();
+
+	(stage as MockStage).simulateMouseDown({ x: 50, y: 50 });
+	(stage as MockStage).simulateMouseMove({ x: 100, y: 100 });
+	(stage as MockStage).simulateMouseUp({ x: 100, y: 100 });
+
+	const config = component.ctx[component.props['config'] as number];
+
+	expect(config).toStrictEqual(oldConfig);
 });
 
 test('sets the correct context', () => {