+++ src/tests/text.test.ts
@@ -18,7 +18,7 @@ test('throws an error if not placed inside a Container (Layer, Group, Label) com
 	expect(() => {
 		render(Text, {
 			props: {
-				config: { x: 0, fontSize: 100, text: 'some text' }
+				config: { x: 0, fontSize: 100, text: 'some text', fill: 'black' }
 			}
 		});
 	}).toThrow(CONTAINER_ERROR);
@@ -28,7 +28,7 @@ test('throws an error if not placed inside a Container (Layer, Group, Label) com
 		render(Text, {
 			context: createMockParentContext(Container.Stage, div),
 			props: {
-				config: { x: 0, fontSize: 100, text: 'some text' }
+				config: { x: 0, fontSize: 100, text: 'some text', fill: 'black' }
 			}
 		});
 	}).toThrow(CONTAINER_ERROR);
@@ -39,7 +39,7 @@ test('is correctly added to the parent Layer', () => {
 	const rendered = render(Text, {
 		context: mockContext,
 		props: {
-			config: { x: 0, fontSize: 100, text: 'some text' }
+			config: { x: 0, fontSize: 100, text: 'some text', fill: 'black' }
 		}
 	});
 
@@ -60,7 +60,7 @@ test('is correctly added to the parent Group', () => {
 	const rendered = render(Text, {
 		context: mockContext,
 		props: {
-			config: { x: 0, fontSize: 100, text: 'some text' }
+			config: { x: 0, fontSize: 100, text: 'some text', fill: 'black' }
 		}
 	});
 
@@ -81,7 +81,7 @@ test('is correctly added to the parent Label', () => {
 	const rendered = render(Text, {
 		context: mockContext,
 		props: {
-			config: { x: 0, fontSize: 100, text: 'some text' }
+			config: { x: 0, fontSize: 100, text: 'some text', fill: 'black' }
 		}
 	});
 
@@ -101,7 +101,7 @@ test('Can listen to Konva events', () => {
 	const rendered = render(Text, {
 		context: createMockParentContext(Container.Layer),
 		props: {
-			config: { x: 0, fontSize: 100, text: 'some text' }
+			config: { x: 0, fontSize: 100, text: 'some text', fill: 'black' }
 		}
 	});
 
@@ -122,7 +122,7 @@ test('Can listen to Konva events', () => {
 });
 
 test('Correctly updates bound config on dragend', () => {
-	const rawConfig = { x: 0, fontSize: 100, text: 'some text' };
+	const rawConfig = { x: 0, fontSize: 100, text: 'some text', fill: 'black' };
 	const CONFIG = { ...rawConfig, draggable: true };
 	const rendered = render(Text, {
 		context: createMockParentContext(Container.Layer),
@@ -148,12 +148,41 @@ test('Correctly updates bound config on dragend', () => {
 	expect(config).toStrictEqual({ ...CONFIG, x: 50 });
 });
 
+test('Does not update config if instantiated with staticConfig prop', () => {
+	const rawConfig = { x: 0, fontSize: 100, text: 'some text', fill: 'black' };
+	const CONFIG = { ...rawConfig, draggable: true };
+	const oldConfig = { ...CONFIG };
+	const rendered = render(Text, {
+		context: createMockParentContext(Container.Layer),
+		props: {
+			config: CONFIG,
+			staticConfig: true
+		}
+	});
+
+	const component = rendered.component.$$;
+	const handle: Konva.Text = component.ctx[component.props['handle'] as number];
+
+	const div = document.createElement('div');
+	const stage = new Konva.Stage({ container: div, width: 1000, height: 1000 });
+
+	stage.add(handle.getLayer()!);
+
+	(stage as MockStage).simulateMouseDown({ x: 50, y: 50 });
+	(stage as MockStage).simulateMouseMove({ x: 100, y: 100 });
+	(stage as MockStage).simulateMouseUp({ x: 100, y: 100 });
+
+	const config = component.ctx[component.props['config'] as number];
+
+	expect(config).toStrictEqual(oldConfig);
+});
+
 test('Does not alter the context', () => {
 	const mockContext = createMockParentContext(Container.Layer);
 	const rendered = render(Text, {
 		context: mockContext,
 		props: {
-			config: { x: 0, fontSize: 100, text: 'some text' }
+			config: { x: 0, fontSize: 100, text: 'some text', fill: 'black' }
 		}
 	});
 
@@ -168,7 +197,7 @@ test('Konva instance is correctly destroyed on component unmount', () => {
 	const rendered = render(Text, {
 		context: mockContext,
 		props: {
-			config: { x: 0, fontSize: 100, text: 'some text' }
+			config: { x: 0, fontSize: 100, text: 'some text', fill: 'black' }
 		}
 	});