+++ src/tests/textpath.test.ts
@@ -148,6 +148,35 @@ test('Correctly updates bound config on dragend', () => {
 	expect(config).toStrictEqual({ ...CONFIG, x: 50 });
 });
 
+test('Does not update config if instantiated with staticConfig prop', () => {
+	const rawConfig = { x: 0, fontSize: 100, text: 'some text', data: 'M 1 60 H 168 Z' };
+	const CONFIG = { ...rawConfig, draggable: true };
+	const oldConfig = { ...CONFIG };
+	const rendered = render(TextPath, {
+		context: createMockParentContext(Container.Layer),
+		props: {
+			config: CONFIG,
+			staticConfig: true
+		}
+	});
+
+	const component = rendered.component.$$;
+	const handle: Konva.TextPath = component.ctx[component.props['handle'] as number];
+
+	const div = document.createElement('div');
+	const stage = new Konva.Stage({ container: div, width: 1000, height: 1000 });
+
+	stage.add(handle.getLayer()!);
+
+	(stage as MockStage).simulateMouseDown({ x: 50, y: 50 });
+	(stage as MockStage).simulateMouseMove({ x: 100, y: 100 });
+	(stage as MockStage).simulateMouseUp({ x: 100, y: 100 });
+
+	const config = component.ctx[component.props['config'] as number];
+
+	expect(config).toStrictEqual(oldConfig);
+});
+
 test('Does not alter the context', () => {
 	const mockContext = createMockParentContext(Container.Layer);
 	const rendered = render(TextPath, {